+++ plugin_ide.ui/src-lang/melnorme/util/swt/components/FieldWidget.java
@@ -14,23 +14,18 @@ import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 
 import melnorme.util.swt.SWTUtil;
-import melnorme.utilbox.core.fntypes.RunnableX;
 import melnorme.utilbox.fields.Field;
 import melnorme.utilbox.fields.FieldValueListener;
 import melnorme.utilbox.fields.FieldValueListener.FieldChangeListener;
-import melnorme.utilbox.fields.validation.IValidatableField;
-import melnorme.utilbox.fields.validation.ValidationField;
-import melnorme.utilbox.fields.validation.ValidationSource;
-import melnorme.utilbox.fields.validation.Validator;
-import melnorme.utilbox.status.StatusException;
 import melnorme.utilbox.fields.IField;
+import melnorme.utilbox.fields.validation.Validator;
 
 /**
  * Field component with a field value that can be manipulated (get/set) even if the 
  * componented is not created.
  */
 public abstract class FieldWidget<VALUE> extends AbstractDisableableWidget 
-	implements IField<VALUE>, IValidatableField<VALUE> {
+	implements IField<VALUE> {
 	
 	private final Field<VALUE> field;
 	
@@ -125,21 +120,9 @@ public abstract class FieldWidget<VALUE> extends AbstractDisableableWidget
 	
 	/* -----------------  ----------------- */
 	
-	@Override
-	public ValidationField getValidation() {
-		return super.getValidation();
-	}
-	
-	public void addFieldValidator(boolean init, Validator<VALUE, ?> validator) {
-		getValidation().addFieldValidator(init, this, validator);
-	}
-	
-	public void addFieldValidation(boolean init, ValidationSource validationSource) {
-		getValidation().addFieldValidation(init, this, validationSource);
-	}
-	
-	public void addFieldValidationX(boolean init, RunnableX<StatusException> validationSource) {
-		getValidation().addFieldValidationX(init, this, validationSource);
+	/*FIXME: */
+	public void addFieldValidator2(boolean init, Validator<VALUE, ?> validator) {
+		getValidation().addFieldValidator2(init, this, validator);
 	}
 	
 }